fix(sniper): Fix rounding error in indentation detection on single type member (#3722)

+++ src/main/java/spoon/support/sniper/internal/IndentationDetector.java
@@ -79,9 +79,9 @@ public class IndentationDetector {
 			indentationSize = 1;
 		}
 
-		boolean usesTabs = wsPrecedingTypeMembers.stream()
+		boolean usesTabs = (double) wsPrecedingTypeMembers.stream()
 				.filter(s -> s.contains("\t"))
-				.count() >= wsPrecedingTypeMembers.size() / 2;
+				.count() >= (double) wsPrecedingTypeMembers.size() / 2;
 		return Pair.of(indentationSize, usesTabs);
 	}
